+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/toolchain/tools/ZigTool.kt
@@ -23,40 +23,18 @@
 package com.falsepattern.zigbrains.project.toolchain.tools
 
 import com.falsepattern.zigbrains.project.toolchain.AbstractZigToolchain
+import com.falsepattern.zigbrains.shared.cli.call
+import com.falsepattern.zigbrains.shared.cli.createCommandLineSafe
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.execution.process.ProcessOutput
-import com.intellij.util.io.awaitExit
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.runInterruptible
-import kotlinx.coroutines.withContext
-import kotlinx.coroutines.withTimeoutOrNull
 import java.nio.file.Path
-import kotlin.io.path.exists
-import kotlin.io.path.isDirectory
-import kotlin.io.path.pathString
 
 abstract class ZigTool(val toolchain: AbstractZigToolchain) {
     abstract val toolName: String
 
     suspend fun callWithArgs(workingDirectory: Path?, vararg parameters: String, timeoutMillis: Long = Long.MAX_VALUE): Result<ProcessOutput> {
         val cli = createBaseCommandLine(workingDirectory, *parameters).let { it.getOrElse { return Result.failure(it) } }
-
-        val (process, exitCode) = withContext(Dispatchers.IO) {
-            val process = cli.createProcess()
-            val exit = withTimeoutOrNull(timeoutMillis) {
-                process.awaitExit()
-            }
-            process to exit
-        }
-        return runInterruptible {
-            Result.success(ProcessOutput(
-                process.inputStream.bufferedReader().use { it.readText() },
-                process.errorStream.bufferedReader().use { it.readText() },
-                exitCode ?: -1,
-                exitCode == null,
-                false
-            ))
-        }
+        return cli.call(timeoutMillis)
     }
 
     private suspend fun createBaseCommandLine(
@@ -64,15 +42,7 @@ abstract class ZigTool(val toolchain: AbstractZigToolchain) {
         vararg parameters: String
     ): Result<GeneralCommandLine> {
         val exe = toolchain.pathToExecutable(toolName)
-        if (!exe.exists())
-            return Result.failure(IllegalArgumentException("file does not exist: ${exe.pathString}"))
-        if (exe.isDirectory())
-            return Result.failure(IllegalArgumentException("file is a directory: ${exe.pathString}"))
-        val cli = GeneralCommandLine()
-            .withExePath(exe.toString())
-            .withWorkingDirectory(workingDirectory)
-            .withParameters(*parameters)
-            .withCharset(Charsets.UTF_8)
-        return Result.success(toolchain.patchCommandLine(cli))
+        return createCommandLineSafe(workingDirectory, exe, *parameters)
+            .mapCatching { toolchain.patchCommandLine(it) }
     }
 }